test(cli): fold findMissingLaunchFiles tests into tests/core/cli.test.ts
CONTRIBUTING L282 prohibits new test files — extend the existing file
that covers the same domain. The 5 findMissingLaunchFiles cases from
the original tests/util/plugin-cache-launch-files.test.ts naturally
belong alongside the existing tests/core/cli.test.ts plugin-cache-
integrity coverage (assertPluginCacheIntegrity, derivePluginManifest,
package.json files[] shipping). All in the same describe block now.

No behaviour change. 5/5 GREEN.
